Former-commit-id: 97b90f4a320c6fac9954b559aae3e047db097a30

+++ server/packet.go
@@ -27,7 +27,7 @@ type Join struct {
 
 type Input struct {
     id string
-    Move int
+    Move float64
     Use int
 }
